+++ src/components/Stages/Levels/index.ts
@@ -1,0 +1,5 @@
+export { Start } from "../../Start/Start";
+export { LevelTwo } from "./LevelTwo";
+export { LevelThree } from "./LevelThree";
+export { LevelOne } from "./LevelOne";
+export { default as FinishClassic } from "../../Finish/FinishClassic";
