+++ src/components/Stages/Quizes/index.ts
@@ -1,0 +1,4 @@
+export { QuizOne } from "../../Start/QuizOne";
+export { QuizTwo } from "./QuizTwo";
+export { QuizThree } from "./QuizThree";
+export { QuizFour } from "./QuizFour";
