+++ src/stories/headers/AppMenu/AppMenu.css
@@ -84,7 +84,7 @@
   transition-duration: 180ms;
 }
 .menu-open-button {
-  z-index: 21;
+  z-index: 2;
   transition-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
   transition-duration: 400ms;
   transform: scale(1.1, 1.1) translate3d(0, 0, 0);
