+++ src/stories/headers/index.ts
@@ -1,2 +1,4 @@
-export { Title } from "./Title/Title";
 export { AppBar } from "./AppBar/AppBar";
+export { AppMenu } from "./AppMenu/AppMenu";
+export { PlaceHolder } from "./PlaceHolder/PlaceHolder";
+export { Title } from "./Title/Title";
